refactor: Restructured, renamed competence scripts for clarity
