+++ index.html
@@ -16,13 +16,13 @@
     <script defer src="https://cloud.umami.is/script.js" data-website-id="64ff51c8-15da-49f3-b671-681e77484dc8"></script>
     <script type="module" src="src/script/scrollAnimation.js"></script>
     <script src="src/script/mouseHighlightEffect.js"></script>
-    <script src="src/script/menuController.js"></script>
     <script type="module" src="src/script/highlightButton.js"></script>
+    <script src="src/script/menuController.js"></script>
     <!--<script src="src/script/cursor.js"></script>-->
-    <script type="module" src="src/script/competence.js"></script>
-    <script type="module" src="src/script/competenceModal.js"></script>
-    <script src="src/script/competenceLoader.js"></script>
-    <script type="module" src="src/script/competenceScroll.js"></script>
+    <script src="src/script/competence/cardLoader.js"></script>
+    <script type="module" src="src/script/competence/scrollController.js"></script>
+    <script type="module" src="src/script/competence/modalController.js"></script>
+    <script type="module" src="src/script/competence/modalLoader.js"></script>
     <script type="module" src="src/script/projectLoader.js"></script>
     <script type="module" src="src/script/projectModal.js"></script>
     <nav class="nav">
