+++ src/script/competence/cardLoader.js
